+++ spec/factories/omses/triggers.rb
@@ -10,5 +10,9 @@ FactoryBot.define do
     factory :trigger_method_get do
       sequence(:method) { :get }
     end
+
+    factory :trigger_with_basic_auth do
+      sequence(:authorization) { build(:authorization_basic) }
+    end
   end
 end