+++ spec/models/oms/authorization_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+RSpec.describe OMS::Authorization, type: :model do
+  it { should belong_to(:trigger) }
+end